+++ app/src/main/java/ryey/easer/plugins/event/sms/SmsEventPlugin.java
@@ -19,18 +19,35 @@
 
 package ryey.easer.plugins.event.sms;
 
+import android.Manifest;
+import android.app.Activity;
 import android.content.Context;
 
 import ryey.easer.commons.plugindef.PluginViewFragment;
 import ryey.easer.commons.plugindef.eventplugin.AbstractSlot;
 import ryey.easer.commons.plugindef.eventplugin.EventData;
 import ryey.easer.commons.plugindef.eventplugin.EventPlugin;
+import ryey.easer.plugins.reusable.PluginHelper;
 
 public class SmsEventPlugin implements EventPlugin {
 
     @Override
     public String name() {
         return "sms";
+    }
+
+    @Override
+    public boolean checkPermissions(Context context) {
+        return PluginHelper.checkPermission(context,
+                Manifest.permission.READ_SMS,
+                Manifest.permission.RECEIVE_SMS);
+    }
+
+    @Override
+    public void requestPermissions(Activity activity, int requestCode) {
+        PluginHelper.requestPermission(activity, requestCode,
+                Manifest.permission.READ_SMS,
+                Manifest.permission.RECEIVE_SMS);
     }
 
     @Override