+++ app/src/main/java/ryey/easer/plugins/operation/command/CommandOperationPlugin.java
@@ -19,6 +19,7 @@
 
 package ryey.easer.plugins.operation.command;
 
+import android.app.Activity;
 import android.content.Context;
 
 import ryey.easer.commons.plugindef.PluginViewFragment;
@@ -36,12 +37,22 @@ public class CommandOperationPlugin implements OperationPlugin {
 
     @Override
     public PrivilegeUsage privilege() {
-        return PrivilegeUsage.no_root;
+        return PrivilegeUsage.prefer_root;
     }
 
     @Override
     public int maxExistence() {
         return -1;
+    }
+
+    @Override
+    public boolean checkPermissions(Context context) {
+        return true;
+    }
+
+    @Override
+    public void requestPermissions(Activity activity, int requestCode) {
+
     }
 
     @Override