+++ app/src/main/java/ryey/easer/plugins/reusable/PluginHelper.java
@@ -1,8 +1,12 @@
 package ryey.easer.plugins.reusable;
 
+import android.app.Activity;
 import android.content.Context;
 import android.content.SharedPreferences;
+import android.content.pm.PackageManager;
 import android.preference.PreferenceManager;
+import android.support.v4.app.ActivityCompat;
+import android.support.v4.content.ContextCompat;
 
 import java.io.IOException;
 
@@ -16,12 +20,22 @@ public class PluginHelper {
                 context.getString(R.string.key_pref_use_root), false);
     }
 
-    public static void executeCommand(Context context, String command) {
-        try {
-            Runtime.getRuntime().exec("su");
-            Runtime.getRuntime().exec(command);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
+    public static Process executeCommandAsRoot(Context context, String command) throws IOException {
+        return Runtime.getRuntime().exec(new String[] { "su", "-c", command});
     }
+
+    public static boolean checkPermission(Context context, String... permissions) {
+        for (String permission : permissions) {
+            if (ContextCompat.checkSelfPermission(context, permission)
+                    != PackageManager.PERMISSION_GRANTED) {
+                return false;
+            }
+        }
+        return true;
+    }
+
+    public static void requestPermission(Activity activity, int requestCode, String... permissions) {
+        ActivityCompat.requestPermissions(activity, permissions, requestCode);
+    }
+
 }